uris-homerobot-core/fine-tuning/requirements.txt
# Dataset Generation Dependencies for Qwen2-VL Fine-tuning
# Python 3.10+

# OpenAI API Client
openai>=1.12.0

# Data Validation
pydantic>=2.6.0

# Token Counting
tiktoken>=0.6.0

# Environment Management
python-dotenv>=1.0.0

# Progress Bar
tqdm>=4.66.0

# Async Support (usually included in Python 3.10+)
asyncio>=3.4.3

# Hugging Face Libraries
transformers>=4.46.0
datasets>=4.4.0
huggingface_hub>=0.36.0
accelerate>=1.12.0
torch>=2.5.0

